Made the `source` property of nodes optional (#19)
This makes the `source` property of all nodes optional. This is a breaking change.

Consequently, the `NoSource` and `Simple` types are no longer necessary and have been removed. `NoParent` is now the sole remaining view for nodes. This caused the signatures of many functions to change as the usages of `Simple<T>` have been replaced with `NoParent<T>`. The `VisitNoSourceAstHandler` and `VisitSimpleAstHandler` interfaces were also removed for obvious reasons.

The `set{Source,Parent}` functions have also been adjusted and no longer return a value. The `desimplify` function has been removed.

---

The main motivations for this are AST transformations and simpler interfaces. Many functions had to support all combinations of `NoParent` and `NoSource` which is no longer necessary. AST transformations are also easier to implement with an optional source because up until now a `source` property was either required or forbidden which makes it very hard to statically type algorithms that can deal with both.